working on feed

+++ blockedDbFunctions.py
@@ -12,6 +12,11 @@ class blockedDB:
         self.connection.row_factory = dict_factory
         self.cursor = self.connection.cursor()
     
+    def checkUserExists(self, id):
+        data = [id, id]
+        self.cursor.execute("SELECT * FROM blocked WHERE blocker_id=? OR blocked_id=?", data)
+        return self.cursor.fetchall()
+    
     def checkBlocked(self, blocked_id, blocker_id):
         data = [blocked_id, blocker_id]
         self.cursor.execute("SELECT 1 FROM blocked WHERE blocked_id=? AND blocker_id=?", data)
@@ -19,13 +24,16 @@ class blockedDB:
 
     # allow one user to block another user
     def insertBlocked(self, blocked_id, blocker_id):
+        if not self.checkUserExists(blocked_id) or not self.checkUserExists(blocker_id):
+            print("Check that both users exist.")
+            return []
         # don't let someone block themselves
         if blocked_id == blocker_id:
-            return
+            return []
         data = [blocked_id, blocker_id]
         if (self.checkBlocked(blocked_id, blocker_id)):
             # print('already blocked')
-            return
+            return []
         # if a user is blocking someone they already follow, or who follows them, when adding the block relationship, delete any follow relationship
         self.cursor.execute("INSERT INTO blocked (blocked_id, blocker_id) VALUES (?, ?)", data)
         self.connection.commit()
@@ -36,6 +44,9 @@ class blockedDB:
 
     # get all users that one person is blocking
     def getAllBlockedByBlockerId(self, blocker_id):
+        if not self.checkUserExists(blocker_id):
+            print("Check that user exists.")
+            return []
         data = [blocker_id]
         self.cursor.execute("SELECT * FROM blocked WHERE blocker_id=?", data)
         users = self.cursor.fetchall()
@@ -43,6 +54,9 @@ class blockedDB:
 
     # allow one user to unblock another user
     def removeBlock(self, blocked_id, blocker_id):
+        if not self.checkUserExists(blocked_id) or not self.checkUserExists(blocker_id):
+            print("Check that both users exist.")
+            return []
         data = [blocked_id, blocker_id]
         self.cursor.execute("DELETE FROM blocked WHERE blocked_id=? AND blocker_id=?", data)
         self.connection.commit()
